feat(renderer): back-to-front depth sorting for the translucent bucket
Translucent draws rendered in submission order — alpha compositing was
only correct if the game happened to submit far-to-near. Each
MaterialDrawCommand now carries its pivot's clip-space w (free at
submit time: mvp[3][3]) and the bucket is stable-sorted back-to-front
before the translucent pass, so equal-depth draws keep submission
order and hand-ordered games render as before. Instanced draws sort as
a group by their fallback-transform pivot.

File policy fallout, resolved by shrinking: material_system's GPU test
module moved to material_system_tests.rs (file now under 2000 —
removed from the grandfather baseline); create_render_texture /
get_texture_ref joined the rest of texture storage in texture_store.rs
(mod.rs 13660 -> 13639).
